Pulled view model logic out of BudgetingController
